\name{bayesELcpp-package}
\alias{bayesELcpp-package}
\alias{bayesELcpp}
\docType{package}
\title{
\packageTitle{bayesELcpp}
}
\description{
\packageDescription{bayesELcpp}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bayesELcpp}
\packageIndices{bayesELcpp}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{bayesELcpp}

Maintainer: \packageMaintainer{bayesELcpp}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
