% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clothianidin}
\alias{clothianidin}
\title{Clothianidin concentration in maize plants.}
\format{
A data frame with 102 rows and 3 variables:
\describe{
  \item{blk}{blocks constructed from original data. Days post planting-blocks-year.}
  \item{trt}{Seed treatments.}
  \item{clo}{log transformed clothianidin concentration in microgram.}
}
}
\source{
Alford A, Krupke CH (2017) Translocation of the neonicotinoid seed treatment clothianidin in maize. PLoS ONE 12(3): e0173836.
\url{https://doi.org/10.1371/journal.pone.0173836}
}
\usage{
data("clothianidin")
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
