% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{confreg,EL-method}
\alias{confreg,EL-method}
\title{Confidence region for model parameters}
\usage{
\S4method{confreg}{EL}(
  object,
  parm,
  level = 0.95,
  cv = qchisq(level, 2L),
  npoints = 50L,
  control = control_el()
)
}
\arguments{
\item{object}{Fitted \code{\link{EL}} object.}

\item{parm}{Specification of which parameters are to be given a confidence
region, either a vector of numbers or a vector of names. It should be a
vector of length two of the form \code{c(x, y)}. If missing, the first two
parameter in \code{object} are considered.}

\item{level}{Confidence level required. Defaults to \code{0.95}.}

\item{cv}{Critical value for calibration of empirical likelihood ratio
statistic. Defaults to \code{qchisq(level, 2L)}.}

\item{npoints}{Number of boundary points to compute. Defaults to \code{50}.}

\item{control}{List of control parameters set by \code{\link{control_el}}.}
}
\value{
An S4 object of class \code{\link{ConfregEL}}.
}
\description{
Computes boundary points of a two-dimensional confidence region for model
  parameters.
}
\examples{
par <- c(0, 0, 0)
x <- matrix(rnorm(90), ncol = 3)
fit <- el_mean2(par, x)
confreg(fit, parm = c(1, 3))
}
\references{
Owen, Art. 1990. “Empirical Likelihood Ratio Confidence Regions.”
  The Annals of Statistics 18 (1): 90–120. \doi{10.1214/aos/1176347494}.
}
\seealso{
\link{confint.el}, \link{control_el}, \link{lht}
}
