% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confreg.R
\name{confreg}
\alias{confreg}
\title{Confidence region for model parameters}
\usage{
confreg(
  object,
  parm,
  level = 0.95,
  cv = qchisq(level, 2L),
  n = 50,
  control = control_el()
)
}
\arguments{
\item{object}{A fitted \code{"el"} object.}

\item{parm}{A specification of which parameters are to be given confidence
intervals, either a vector of numbers or a vector of names. If missing, all
parameters are considered.}

\item{level}{A confidence level required.}

\item{cv}{A critical value for calibration of empirical likelihood ratio
statistic. Defaults to \code{qchisq(level, 2L)}.}

\item{control}{A list of control parameters set by \code{\link{control_el}}.}
}
\value{
A matrix with columns giving lower and upper confidence limits for
 each parameter. In contrast to other methods that rely on studentization,
 the lower and upper limits obtained from empirical likelihood do not
 correspond to the \code{(1 - level) / 2} and \code{1 - (1 - level) / 2} in
 \%, respectively.
}
\description{
Computes confidence region for two model parameters in a fitted model.
}
\examples{
fit <- el_lm(formula = mpg ~ wt, data = mtcars)
confint(fit)
}
\references{
Owen, Art. 1990. “Empirical Likelihood Ratio Confidence Regions.”
  The Annals of Statistics 18 (1): 90–120. \doi{10.1214/aos/1176347494}.
}
\seealso{
\link{control_el}, \link{confint.el}
}
