% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ConfregEL-class}
\alias{ConfregEL-class}
\alias{ConfregEL}
\title{S4 class \code{ConfregEL}}
\description{
S4 class for confidence region.
}
\section{Slots}{

\describe{
\item{\code{points}}{Numeric matrix with two columns for boundary points of a
confidence region.}

\item{\code{estimates}}{Numeric vector of length two for parameter estimates.}

\item{\code{level}}{Confidence level required.}

\item{\code{cv}}{Critical value for calibration of empirical likelihood ratio
statistic.}

\item{\code{pnames}}{Character vector of length two for the name of parameters.}
}}

