% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{EL-class}
\alias{EL-class}
\title{S4 class \linkS4class{EL}}
\description{
S4 class for empirical likelihood.
}
\section{Slots}{

\describe{
\item{\code{optim}}{A list with the following optimization results:
\itemize{
\item{\code{method } }{Character for method dispatch in internal
functions.}
\item{\code{par } }{Value at which empirical likelihood is evaluated.}
\item{\code{lambda } }{Lagrange multiplier of the dual problem.}
\item{\code{iterations } }{Number of iterations performed.}
\item{\code{convergence } }{Convergence status.}
}}

\item{\code{logp}}{Log probabilities obtained from empirical likelihood.}

\item{\code{logl}}{Empirical log-likelihood.}

\item{\code{loglr}}{Empirical log-likelihood ratio.}

\item{\code{statistic}}{Minus twice the empirical log-likelihood ratio.}

\item{\code{df}}{Degrees of freedom of the statistic.}

\item{\code{pval}}{\eqn{p}-value of the statistic.}

\item{\code{npar}}{Number of parameters.}

\item{\code{weights}}{Rescaled weights used for model fitting.}

\item{\code{dataMatrix}}{Data matrix used for model fitting.}

\item{\code{coefficients}}{Maximum empirical likelihood estimates of the parameters.}
}}

\examples{
showClass("EL")
}
