% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/el.mean.R
\name{el.mean}
\alias{el.mean}
\title{Compute empirical likelihood for mean}
\usage{
el.mean(theta, x, abstol = 1e-08, maxit = 50)
}
\arguments{
\item{theta}{a vector of parameters to be tested.}

\item{x}{a matrix or vector of data. Each row is an observation vector.}

\item{abstol}{an optional value for the absolute convergence tolerance. Defaults to 1e-8.}

\item{maxit}{an optional value for the maximum number of iterations. Defaults to 50.}
}
\value{
An object with S3 class "el"
}
\description{
Compute empirical likelihood for mean
}
\examples{
theta <- 0
x <- rnorm(100)
el.mean(theta, x)
}
