% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/el_aov.R
\name{el_aov}
\alias{el_aov}
\title{Empirical likelihood pairwise comparisons}
\usage{
el_aov(formula, data, maxit = 10000, abstol = 1e-08)
}
\arguments{
\item{formula}{A formula object. It must specify variables for response and treatment as 'response ~ treatment'.}

\item{data}{A data frame containing the variables in the formula.}

\item{maxit}{Maximum number of iterations for optimization. Defaults to 10000.}

\item{abstol}{Absolute convergence tolerance for optimization. Defaults to 1e-08.}
}
\value{
A list with class \code{c("el_aov", "melt")}.
}
\description{
Tests all pairwise comparisons or comparisons with control for general block designs. Two single step asymptotic \eqn{k}-FWER (generalized family-wise error rate) controlling procedures are available: asymptotic Monte Carlo (AMC) and nonparametric bootstrap (NB).
}
\examples{
el_aov(clo ~ trt, clothianidin)

}
