% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/el_control.R
\name{el_control}
\alias{el_control}
\title{Control parameters for computation}
\usage{
el_control(
  maxit = 200L,
  maxit_l = 25L,
  tol = 1e-06,
  tol_l = 1e-06,
  step = NULL,
  th = NULL,
  nthreads,
  seed = sample.int(.Machine$integer.max, 1L),
  B = 10000L
)
}
\arguments{
\item{maxit}{A single integer for the maximum number of iterations for
constrained minimization of empirical likelihood. Defaults to \code{200}.}

\item{maxit_l}{A single integer for the maximum number of iterations for
evaluation of empirical likelihood. Defaults to \code{25}.}

\item{tol}{A single numeric for the convergence tolerance for the constrained
minimization. Defaults to \code{1e-06}.}

\item{tol_l}{A single numeric for the relative convergence tolerance for the
evaluation. Defaults to \code{1e-06}.}

\item{step}{A single numeric for the step size for projected gradient descent
method. Defaults to \code{NULL} and set to the reciprocal of sample size.}

\item{th}{A single numeric for the threshold for negative empirical
log-likelihood ratio value. The iteration stops if the value exceeds the
threshold. Defaults to \code{NULL}.}

\item{nthreads}{A single integer for the number of threads for parallel
computation via OpenMP (if available).}

\item{seed}{A single integer for the number of .}

\item{B}{A single integer for the number of . Defaults to \code{10000L}.}
}
\value{
An object of class of \linkS4class{ControlEL}.
}
\description{
Specifies details of computation of (constrained) empirical likelihood.
}
\examples{
optcfg <- el_control(maxit = 300L, th = 200, nthreads = 1L)
}
\seealso{
\link{el_eval}, \link{elt}
}
