% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/el_eval.R
\name{el_eval}
\alias{el_eval}
\title{Empirical likelihood with general estimating functions}
\usage{
el_eval(g, weights = NULL, control = list())
}
\arguments{
\item{g}{A numeric matrix, or an object that can be coerced to a numeric matrix. Each row corresponds to an observation.}

\item{weights}{An optional numeric vector of weights. Defaults to \code{NULL}, corresponding to identical weights. If non-\code{NULL}, weighted empirical likelihood is computed.}

\item{control}{A list of control parameters. See ‘Details’.}
}
\value{
A list with class \code{"el_test"}.
}
\description{
Computes empirical likelihood with general estimating functions.
}
