% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/el_glm.R
\name{el_glm}
\alias{el_glm}
\title{Empirical likelihood for generalized linear models}
\usage{
el_glm(
  formula,
  family = gaussian,
  data,
  weights = NULL,
  na.action,
  control = control_el(),
  model = TRUE,
  start = NULL,
  etastart = NULL,
  mustart = NULL,
  ...
)
}
\arguments{
\item{formula}{An object of class \code{"\link[stats]{formula}"} (or one that
can be coerced to that class): a symbolic description of the model to be
fitted.}

\item{family}{A description of the error distribution and link function to be
used in the model. Only the result of a call to a family function is
supported. See ‘Details’.}

\item{data}{An optional data frame, list or environment (or object coercible
by \code{\link[base]{as.data.frame}} to a data frame) containing the
variables in the formula. If not found in data, the variables are taken
from \code{environment(formula)}.}

\item{weights}{An optional numeric vector of weights to be used in the
fitting process. Defaults to \code{NULL}, corresponding to identical
weights. If non-\code{NULL}, weighted empirical likelihood is computed.
See ‘Details’.}

\item{na.action}{A function which indicates what should happen when the data
contain \code{NA}s. The default is set by the \code{na.action} setting of
\code{\link[base]{options}}, and is \code{na.fail} if that is unset.}

\item{control}{A list of control parameters set by
\code{\link{control_el}}.}

\item{model}{A logical. If \code{TRUE} the data matrix used for fitting is
returned.}

\item{start}{Starting values for the parameters in the linear predictor.
Defaults to \code{NULL} and is passed to \code{\link[stats]{glm.fit}}.}

\item{etastart}{Starting values for the linear predictor. Defaults to
\code{NULL} and is passed to \code{\link[stats]{glm.fit}}.}

\item{mustart}{Starting values for the vector of means. Defaults to
\code{NULL} and is passed to \code{\link[stats]{glm.fit}}.}

\item{...}{Additional arguments to be passed to
\code{\link[stats]{glm.control}}.}
}
\value{
A list of class \code{c("el_glm", "el_lm", "el")} with the following
  components:
  \item{optim}{A list with the following optimization results:
    \itemize{
      \item{\code{method } }{A character for method dispatch in internal
      functions.}
      \item{\code{par } }{The solution of the constrained minimization
      problem.}
      \item{\code{lambda } }{The Lagrange multiplier of dual problem.}
      \item{\code{logLR } }{The (weighted) empirical log-likelihood ratio
      value.}
      \item{\code{iterations } }{The number of iterations performed.}
      \item{\code{convergence } }{A logical vector. \code{TRUE} indicates
      convergence of the algorithm.}
    }
  }
  \item{par.tests}{A list with the test results for each parameter:
    \itemize{
      \item{\code{statistic } }{A numeric vector of chi-squared statistics.}
      \item{\code{convergence } }{A logical vector. \code{TRUE} indicates
      convergence of the algorithm.}
    }
  }
  \item{log.prob}{The log probabilities.}
  \item{loglik}{The log likelihood value.}
  \item{statistic}{The chi-square statistic.}
  \item{df}{The degrees of freedom of the statistic.}
  \item{p.value}{The \eqn{p}-value of the statistic.}
  \item{npar}{The number of parameters.}
  \item{weights}{The rescaled weights if non-\code{NULL} \code{weights} is
  supplied}
  \item{data.matrix}{The data matrix used for fitting if \code{model} is
  \code{TRUE}.}
  \item{coefficients}{The maximum empirical likelihood estimates of the
  parameters.}
}
\description{
Fits a generalized linear model with empirical likelihood.
}
\examples{
n <- 50
x <- rnorm(n)
x2 <- rnorm(n)
l <- -2 + 0.2 * x + 3 * x2
mu <- 1 / (1 + exp(-l))
y <- rbinom(n, 1, mu)
df <- data.frame(y, x, x2)
fit <- el_glm(y ~ x + x2, family = binomial, df)
summary(fit)
}
\references{
Chen, Song Xi, and Hengjian Cui. 2003.
  “An Extended Empirical Likelihood for Generalized Linear Models.”
  Statistica Sinica 13: 69–81.
}
\seealso{
\link{el_lm}, \link{control_el}, \link{lht}
}
