% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/el_mean2.R
\name{el_mean2}
\alias{el_mean2}
\title{Empirical likelihood for the mean}
\usage{
el_mean2(par, x, weights = NULL, control = control_el(), model = TRUE)
}
\arguments{
\item{par}{Numeric vector of parameter values to be tested.}

\item{x}{Numeric matrix, or an object that can be coerced to a numeric
matrix. Each row corresponds to an observation.}

\item{weights}{Optional numeric vector of weights to be used in the fitting
process. Defaults to \code{NULL}, corresponding to identical weights. If
non-\code{NULL}, weighted empirical likelihood is computed.}

\item{control}{List of control parameters set by \code{\link{control_el}}.}

\item{model}{Logical. If \code{TRUE} the data matrix used for model fitting
is returned.}
}
\value{
An S4 object of class \linkS4class{EL}.
}
\description{
Computes empirical likelihood for the mean.
}
\examples{
# scalar mean
par <- 0
x <- rnorm(100)
el_mean2(par, x)

# vector mean
par <- c(0, 0)
x <- matrix(rnorm(100), ncol = 2)
el_mean2(par, x)

# weighted data
par <- c(0, 0)
x <- matrix(rnorm(100), ncol = 2)
w <- rep(c(1, 2), each = 25)
el_mean2(par, x, w)
}
\references{
Glenn, N.L., and Yichuan Zhao. 2007.
  “Weighted Empirical Likelihood Estimates and Their Robustness Properties.”
  Computational Statistics & Data Analysis 51 (10): 5130–41.
  \doi{10.1016/j.csda.2006.07.032}.

Owen, Art. 1990. “Empirical Likelihood Ratio Confidence Regions.”
  The Annals of Statistics 18 (1): 90–120. \doi{10.1214/aos/1176347494}.
}
\seealso{
\link{control_el}, \link{el_eval}, \link{lht}
}
