% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{el_pairwise}
\alias{el_pairwise}
\title{Pairwise Comparisons for General Block Design}
\usage{
el_pairwise(
  x,
  c,
  control = 0L,
  k = 1L,
  level = 0.05,
  interval = TRUE,
  method = "AMC",
  B = 10000L,
  approx = FALSE,
  nthread = 1L,
  maxit = 10000L,
  abstol = 1e-08
)
}
\arguments{
\item{x}{a matrix of data .}

\item{c}{an incidence matrix.}

\item{control}{control treatment. Defaults to 0.}

\item{k}{integer k for k-FWER. Defaults to 1.}

\item{level}{level.}

\item{interval}{whether to compute interval. Defaults to TRUE.}

\item{method}{the method to be used; either 'AMC' or 'NB' is supported. Defaults to 'AMC'.}

\item{B}{number of bootstrap replicates.}

\item{approx}{whether to use the approximation for lambda. Defaults to FALSE.}

\item{nthread}{number of cores(threads) to use. Defaults to 1.}

\item{maxit}{an optional value for the maximum number of iterations. Defaults to 1000.}

\item{abstol}{an optional value for the absolute convergence tolerance. Defaults to 1e-8.}
}
\description{
Either all pairwise comparisons or comparisons with control is available.
}
