% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/el_test.R
\name{el_test2}
\alias{el_test2}
\title{Test2 (development version)}
\usage{
el_test2(object, rhs, control = list())
}
\arguments{
\item{object}{An object}

\item{rhs}{A rhs}

\item{control}{A list of control parameters. See ‘Details’ in
\code{\link{el_eval}}.}
}
\description{
Test2 (development version)
}
