% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ELD-class}
\alias{ELD-class}
\alias{ELD}
\title{S4 class \code{ELD}}
\description{
S4 class for empirical likelihood displacement.
}
\section{Slots}{

\describe{
\item{\code{eld}}{Numeric vector of empirical likelihood displacement values.}
}}

