% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eld.R
\name{eld}
\alias{eld}
\title{Empirical likelihood displacement}
\usage{
eld(object, control = list())
}
\arguments{
\item{object}{A fitted \code{"el_test"} object.}

\item{control}{A list of control parameters. See ‘Details’.}
}
\description{
Empirical likelihood displacement for objects inheriting from class
  \code{"el_test"}.
}
