% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELmht.R
\name{ELmht}
\alias{ELmht}
\title{Empirical Likelihood Multiple Hypothesis Testing}
\usage{
ELmht(
  df,
  hypotheses,
  rhs = NULL,
  k = 1,
  level = 0.05,
  method = c("AMC", "NB"),
  B = 10000,
  approx = F,
  nthread = 1,
  maxit = 10000,
  abstol = 1e-08
)
}
\arguments{
\item{df}{A data frame.}

\item{hypotheses}{hypotheses.}

\item{rhs}{Right hand side of hypotheses}

\item{k}{Integer k for k-FWER. Defaults to 1.}

\item{level}{Level of test.}

\item{method}{Method to be used; either `AMC' or `NB' is supported. Defaults to `AMC'.}

\item{B}{Number of replicates for AMC and NB.}

\item{approx}{Whether to use the approximation for NB. Defaults to FALSE.}

\item{nthread}{Number of cores(threads) to use for NB. Defaults to 1.}

\item{maxit}{Maximum number of iterations for optimization. Defaults to 1000.}

\item{abstol}{Absolute convergence tolerance. Defaults to 1e-08.}
}
\value{
A list with class \code{elmulttest}.
}
\description{
Empirical Likelihood Multiple Hypothesis Testing
}
\examples{
sum(1:10)
sum(1:5, 6:10)

}
