% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ELMT-class}
\alias{ELMT-class}
\title{S4 class \linkS4class{ELMT}}
\description{
S4 class for multiple empirical likelihood test.
}
\section{Slots}{

\describe{
\item{\code{statistic}}{A single numeric for the minus twice the (constrained)
empirical log-likelihood ratio.}
}}

\examples{
showClass("ELMT")
}
