% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/print-methods.R
\docType{class}
\name{ELT-class}
\alias{ELT-class}
\alias{print,ELT-method}
\title{S4 class \linkS4class{ELT}}
\usage{
\S4method{print}{ELT}(x, digits = getOption("digits"), ...)
}
\description{
S4 class for empirical likelihood test.
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: prints test results
}}

\section{Slots}{

\describe{
\item{\code{optim}}{A list with the optimization results.}

\item{\code{alpha}}{A single numeric for the significance level.}

\item{\code{logl}}{A single numeric for the (constrained) empirical log-likelihood.}

\item{\code{statistic}}{A single numeric for the minus twice the (constrained)
empirical log-likelihood ratio.}

\item{\code{cv}}{A single numeric for the critical value.}

\item{\code{pval}}{A single numeric for the \eqn{p}-value of the statistic.}

\item{\code{calibrate}}{A single character for the calibration method used.}
}}

\examples{
showClass("ELT")
}
