% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/el_lm.R
\name{formula.el_lm}
\alias{formula.el_lm}
\title{formula}
\usage{
\method{formula}{el_lm}(x, ...)
}
\arguments{
\item{x}{object of class "el_lm".}

\item{...}{further arguments to be passed to or from methods.}
}
\description{
formula
}
