% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getOptim-methods.R
\name{getOptim}
\alias{getOptim}
\alias{getOptim,EL-method}
\alias{getOptim,ELT-method}
\title{Optimization results}
\usage{
\S4method{getOptim}{EL}(object, ...)

\S4method{getOptim}{ELT}(object, ...)
}
\arguments{
\item{object}{An object that inherit from \linkS4class{EL} or
\linkS4class{ELT}.}

\item{...}{Further arguments passed to methods.}
}
\value{
A list.
}
\description{
Extracts optimization results from a model.
}
\examples{
data("precip")
fit <- el_mean(precip, par = 40)
getOptim(fit)
}
