% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getSigTests-methods.R
\name{getSigTests}
\alias{getSigTests}
\alias{getSigTests,LM-method}
\title{Significance tests}
\usage{
\S4method{getSigTests}{LM}(object)
}
\arguments{
\item{object}{An object that inherit from \linkS4class{LM}.}
}
\value{
A list with the following optimization results:
\itemize{
\item \code{statistic} A numeric vector of the minus twice the empirical
log-likelihood ratio statistics.
\item \code{iterations} An integer vector of the number of iterations performed.
\item \code{convergence} A logical vector of the convergence statuses.
}
}
\description{
Extracts the results of significance tests from a model.
}
\examples{
data("mtcars")
fit <- el_lm(mpg ~ ., data = mtcars)
getSigTests(fit)
}
