% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{GLM-class}
\alias{GLM-class}
\alias{GLM}
\title{\linkS4class{GLM} class}
\description{
S4 class for generalized linear models. It inherits from \linkS4class{LM}
class.
}
\section{Slots}{

\describe{
\item{\code{family}}{A \code{\link{family}} object used.}

\item{\code{dispersion}}{A single numeric for the estimated dispersion parameter.}

\item{\code{sigTests}}{A list with the results of significance tests.}

\item{\code{call}}{A matched call.}

\item{\code{terms}}{A \code{\link{terms}} object used.}

\item{\code{misc}}{A list with miscellaneous outputs from a model fitting function.
They are used in other generics and methods.}

\item{\code{optim}}{A list with the following optimization results:
\itemize{
\item \code{par} A numeric vector of the specified parameters.
\item \code{lambda} A numeric vector of the Lagrange multipliers.
\item \code{iterations} A single integer for the number of iterations performed.
\item \code{convergence} A single logical for the convergence status.
}}

\item{\code{logp}}{A numeric vector of the log probabilities obtained from empirical
likelihood.}

\item{\code{logl}}{A single numeric for the empirical log-likelihood.}

\item{\code{loglr}}{A single numeric for the empirical log-likelihood ratio.}

\item{\code{statistic}}{A single numeric for the minus twice the empirical
log-likelihood ratio statistic that has an asymptotic chi-square
distribution.}

\item{\code{df}}{A single integer for the degrees of freedom of the statistic.}

\item{\code{pval}}{A single numeric for the \eqn{p}-value of the statistic.}

\item{\code{nobs}}{A sinble integer for the number of observations.}

\item{\code{npar}}{A single integer for the number of parameters.}

\item{\code{weights}}{A numeric vector of re-scaled weights used for model fitting.}

\item{\code{coefficients}}{A numeric vector of the maximum empirical likelihood
estimates of the parameters.}

\item{\code{method}}{A single character for the method dispatch in internal
functions.}

\item{\code{data}}{A numeric matrix for the data used for model fitting.}
}}

\examples{
showClass("GLM")
}
