% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik-methods.R
\name{logLik,EL-method}
\alias{logLik,EL-method}
\title{Empirical log-likelihood}
\usage{
\S4method{logLik}{EL}(object, ...)
}
\arguments{
\item{object}{A fitted \linkS4class{EL} object.}

\item{...}{Some methods for this generic function require extra arguments.
None are used in this method.}
}
\value{
An object of class \code{"logLik"} with an attribute \code{df} that
  gives the number of (estimated) parameters in the model.
}
\description{
Extracts empirical log-likelihood of a model represented evaluated at the
  estimated coefficients. Package \strong{melt} adds a method for objects
  inheriting from class \linkS4class{EL}.
}
\examples{
fit <- el_lm(formula = mpg ~ wt, data = mtcars)
logLik(fit)
}
