% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{logLik.el}
\alias{logLik.el}
\title{Empirical log-likelihood}
\usage{
\method{logLik}{el}(object, ...)
}
\arguments{
\item{object}{A fitted \code{"el"} object.}

\item{...}{Some methods for this generic function require extra arguments.
None are used in this method.}
}
\value{
An object of class \code{"logLik"} with an attribute \code{df} that
  gives the number of (estimated) parameters in the model.
}
\description{
Extracts empirical log-likelihood of a model represented by
  \code{object} evaluated at the estimated coefficients. Package
  \strong{melt} adds a method for objects inheriting from class \code{"el"}.
}
\examples{
fit <- el_lm(formula = mpg ~ wt, data = mtcars)
logLik(fit)
}
