% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/el_lm.R
\name{logLik.el_lm}
\alias{logLik.el_lm}
\title{logLik}
\usage{
\method{logLik}{el_lm}(object, ...)
}
\arguments{
\item{object}{object of class "el_lm".}

\item{...}{further arguments to be passed to or from methods.}
}
\description{
logLik
}
