% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/el_test.R
\name{logLik.el_test}
\alias{logLik.el_test}
\title{Empirical log-likelihood}
\usage{
\method{logLik}{el_test}(object, ...)
}
\arguments{
\item{object}{A fitted \code{"el_test"} object.}

\item{...}{Some methods for this generic function require extra arguments.
None are used in this method.}
}
\value{
An object of class \code{"logLik"} with an attribute \code{df} that
  gives the number of (estimated) parameters in the model.
}
\description{
Computes the empirical log-likelihood value of the model represented by
  \code{object} evaluated at the estimated coefficients. Package
  \strong{melt} adds a method for objects inheriting from class
  \code{"el_test"}.
}
\examples{
fit <- el_lm(formula = mpg ~ wt, data = mtcars)
logLik(fit)
}
