% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/print-methods.R
\docType{class}
\name{logLikEL-class}
\alias{logLikEL-class}
\alias{logLikEL}
\alias{print,logLikEL-method}
\title{S4 class \linkS4class{logLikEL}}
\usage{
\S4method{print}{logLikEL}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{x}{An object of class \linkS4class{logLikEL}.}

\item{digits}{A single integer for the number of significant digits to be
passed to \code{\link[=format]{format()}}.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
S4 class for empirical log-likelihood.
}
\section{Slots}{

\describe{
\item{\code{logLik}}{A single numeric for the empirical log-likelihood.}

\item{\code{df}}{A single integer for the degrees of freedom or the number of
(estimated) parameters in the model.}
}}

\examples{
showClass("logLikEL")
}
