% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/logLR-methods.R
\name{logLR}
\alias{logLR}
\alias{logLR,EL-method}
\title{Empirical log-likelihood ratio}
\usage{
\S4method{logLR}{EL}(object, ...)
}
\arguments{
\item{object}{An object that inherit from class \linkS4class{EL}, including
\linkS4class{CEL}, \linkS4class{LM}, and \linkS4class{GLM}.}

\item{...}{Further arguments passed to other methods.}
}
\value{
A single numeric.
}
\description{
Extracts empirical log-likelihood ratio from a model.
}
\examples{
fit <- el_lm(mpg ~ wt, data = mtcars)
logLR(fit)
}
