% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/logProb-methods.R
\name{logProb}
\alias{logProb}
\alias{logProb,EL-method}
\alias{logProb,ELT-method}
\title{Log probabilities}
\usage{
\S4method{logProb}{EL}(object, ...)

\S4method{logProb}{ELT}(object, ...)
}
\arguments{
\item{object}{An object that inherits from \linkS4class{EL} or
\linkS4class{ELT}.}

\item{...}{Further arguments passed to methods.}
}
\value{
A numeric vector.
}
\description{
Extracts log probabilities of empirical likelihood from a model.
}
\examples{
data("precip")
fit <- el_mean(precip, par = 40)
logProb(fit)
}
\seealso{
\linkS4class{EL}, \linkS4class{ELT}
}
