% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/nobs-methods.R
\name{nobs}
\alias{nobs}
\alias{nobs,EL-method}
\title{Number of observations in a model}
\usage{
nobs(object, ...)

\S4method{nobs}{EL}(object, ...)
}
\arguments{
\item{object}{An object that inherit from class \linkS4class{EL}, including
\linkS4class{CEL}, \linkS4class{LM}, and \linkS4class{GLM}.}

\item{...}{Further arguments passed to other methods.}
}
\value{
A single integer.
}
\description{
Extracts the number of observations from a model.
}
\examples{
fit <- el_lm(mpg ~ wt, data = mtcars)
nobs(fit)
}
