% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pairwise_PB_ibd}
\alias{pairwise_PB_ibd}
\title{Pairwise comparison for BIBD}
\usage{
pairwise_PB_ibd(
  x,
  c,
  interval = FALSE,
  B = 100000L,
  level = 0.05,
  vcov_adj = FALSE,
  maxit = 1000L,
  abstol = 1e-08
)
}
\arguments{
\item{x}{a matrix of data .}

\item{c}{an incidence matrix.}

\item{interval}{whether to compute interval. Defaults to FALSE.}

\item{B}{number of bootstrap replicate.}

\item{level}{confidence level.}

\item{vcov_adj}{whether to adjust for the covariance estimate. Defaults to FALSE.}

\item{maxit}{an optional value for the maximum number of iterations. Defaults to 1000.}

\item{abstol}{an optional value for the absolute convergence tolerance. Defaults to 1e-8.}
}
\description{
Pairwise comparison for BIBD
}
