% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pairwise_threshold}
\alias{pairwise_threshold}
\title{Computes threshold of pairwise comparison for BIBD}
\usage{
pairwise_threshold(x, c, B, alpha)
}
\arguments{
\item{x}{a matrix of data .}

\item{c}{an incidence matrix.}

\item{B}{an integer value for the number of bootstrap replicates.}

\item{alpha}{a significance level.}
}
\description{
Computes threshold using parametric bootstrap
}
