% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{plot-method}
\alias{plot-method}
\alias{plot}
\alias{plot,ConfregEL-method}
\alias{plot,ELD-method}
\title{Plot methods for package \strong{melt}}
\usage{
plot(x, y, ...)

\S4method{plot}{ConfregEL}(x, y, tt = NULL, ...)

\S4method{plot}{ELD}(
  x,
  y,
  ...,
  main = "Empirical Likelihood Displacement",
  ylab = "ELD",
  pch = 21
)
}
\arguments{
\item{x}{An S4 object of class \code{\link{ConfregEL}} or \code{\link{ELD}}.}

\item{y}{y From the generic \code{plot} function, ignored for class objects.}

\item{...}{A character vector of length two for the name of parameters.}
}
\description{
Provides plot methods for S4 objects that inherit from class \code{"el"}.
}
\examples{
data("clothianidin")
fit <- el_lm(clo ~ trt, clothianidin)
eld <- eld(fit)
plot(eld)
}
\seealso{
\link{confreg}, \link{eld}
}
