% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eld.R
\name{plot.eld}
\alias{plot.eld}
\title{Plot for eld objects}
\usage{
\method{plot}{eld}(
  x,
  ...,
  main = "Empirical Likelihood Displacement",
  ylab = "ELD",
  pch = 21
)
}
\arguments{
\item{x}{An object of class \code{`eld`}.}

\item{...}{Additional arguments to be passed to \code{\link[base]{plot}}.}

\item{main}{A title for the plot.}

\item{ylab}{A label for the y-axis.}

\item{pch}{A vector of plotting characters or symbols to use.}
}
\description{
Takes a fitted \code{`eld`} object and plots the empirical likelihood
  displacement values versus the observation index.
}
\examples{
data("clothianidin")
fit <- el_lm(clo ~ trt, clothianidin)
eld <- eld(fit)
plot(eld)
}
\seealso{
\link{eld}
}
