% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plot-methods.R
\name{plot}
\alias{plot}
\alias{plot,ConfregEL-method}
\alias{plot,ELD-method}
\title{Plot methods}
\usage{
\S4method{plot}{ConfregEL}(x, y, ...)

\S4method{plot}{ELD}(x, y, ...)
}
\arguments{
\item{x}{An object to be plotted.}

\item{y}{Not used.}

\item{...}{Further graphical parameters (see \code{\link[graphics]{par}}).}
}
\description{
Provides plot methods for objects that inherit from class \linkS4class{EL}.
}
\section{Methods (by class)}{
\itemize{
\item \code{ConfregEL}: Plots a two-dimensional confidence region for model
parameters.

\item \code{ELD}: Plots empirical likelihood displacement values versus
observation index.
}}

\examples{
# model
fit <- el_lm(hp ~ wt, data = mtcars)

# confidence region
out1 <- confreg(fit, npoints = 500)
plot(out1)

# empirical likelihood displacement
out2 <- eld(fit)
plot(out2)
}
\seealso{
\link{confreg}, \link{eld}
}
