% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/el_test.R
\name{print.el_test}
\alias{print.el_test}
\title{Prints surveyor object.}
\usage{
\method{print}{el_test}(x, digits = getOption("digits"), prefix = "\\t", ...)
}
\arguments{
\item{x}{object of class "el_test".}

\item{digits}{number of significant digits to be used.}

\item{prefix}{string, passed to strwrap for displaying the method component of the htest object.}

\item{...further}{arguments to be passed to or from methods.}
}
\description{
Prints surveyor object.
}
