% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/print-methods.R
\name{print}
\alias{print}
\alias{print,EL-method}
\alias{print,SummaryLM-method}
\alias{print,logLikEL-method}
\alias{print,ELT-method}
\title{Print methods}
\usage{
\S4method{print}{EL}(x, digits = max(3L, getOption("digits") - 3L), ...)

\S4method{print}{SummaryLM}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"),
  ...
)

\S4method{print}{logLikEL}(x, digits = getOption("digits"), ...)

\S4method{print}{ELT}(x, digits = getOption("digits"), ...)
}
\arguments{
\item{x}{An object to be printed.}

\item{...}{Further arguments passed to other methods.}

\item{digits}{The number of significant digits to be passed to
\code{\link[base]{format}}.}

\item{signif.stars}{A single logical. If \code{TRUE}, ‘significance stars’
are printed for each coefficient.}
}
\description{
Provides print methods for the objects from the package \pkg{melt}.
}
