% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/pval-methods.R
\name{pval}
\alias{pval}
\alias{pval,EL-method}
\alias{pval,ELT-method}
\alias{pval,ELMT-method}
\title{\eqn{p}-value methods}
\usage{
\S4method{pval}{EL}(object, ...)

\S4method{pval}{ELT}(object, ...)

\S4method{pval}{ELMT}(object, ...)
}
\arguments{
\item{object}{An object that inherit from \linkS4class{EL},
\linkS4class{ELT}, or \linkS4class{ELMT}.}

\item{...}{Further arguments passed to methods.}
}
\value{
The form of the value returned by \code{\link[=pval]{pval()}} depends on the class of
its argument.
}
\description{
Extracts the \eqn{p}-value from a model.
}
\section{Methods (by class)}{
\itemize{
\item \code{pval(EL)}: Extracts the \eqn{p}-value.

\item \code{pval(ELT)}: Extracts the \eqn{p}-value.

\item \code{pval(ELMT)}: Extracts the multiplicity adjusted \eqn{p}-values.

}}
\examples{
data("precip")
fit <- el_mean(precip, par = 40)
pval(fit)
}
