% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{QGLM-class}
\alias{QGLM-class}
\alias{QGLM}
\title{\linkS4class{QGLM} class}
\description{
S4 class for generalized linear models with quasi-likelihood methods. It
inherits from \linkS4class{GLM} class.
}
\examples{
showClass("QGLM")
}
