% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/sigTests-methods.R
\name{sigTests}
\alias{sigTests}
\alias{sigTests,LM-method}
\alias{sigTests,SummaryLM-method}
\title{Significance tests}
\usage{
\S4method{sigTests}{LM}(object, ...)

\S4method{sigTests}{SummaryLM}(object, ...)
}
\arguments{
\item{object}{An object that inherits from \linkS4class{LM} or
\linkS4class{SummaryLM}.}

\item{...}{Further arguments passed to methods.}
}
\value{
The form of the value returned by \code{\link[=sigTests]{sigTests()}} depends on the
class of its argument.
}
\description{
Extracts the results of significance tests from a model.
}
\section{Methods (by class)}{
\itemize{
\item \code{sigTests(LM)}: Extracts a list with the optimization results of
significance tests.

\item \code{sigTests(SummaryLM)}: Extracts a matrix with the results of significance
tests.

}}
\examples{
data("mtcars")
fit <- el_lm(mpg ~ ., data = mtcars)
sigTests(fit)
sigTests(summary(fit))
}
\seealso{
\linkS4class{LM}, \linkS4class{SummaryLM}, \code{\link[=getOptim]{getOptim()}}
}
