% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/summary-methods.R
\name{summary-method}
\alias{summary-method}
\alias{summary,LM-method}
\title{Summary methods}
\usage{
\S4method{summary}{LM}(object, ...)
}
\arguments{
\item{object}{An object for which a summary is desired.}

\item{...}{Additional arguments affecting the summary produced.}
}
\description{
Provides summary methods for objects that inherit from class
  \linkS4class{EL}.
}
\section{Functions}{
\itemize{
\item \code{summary,LM-method}: Summarizes the results of the overall test and the
tests for each parameter.
}}

