% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SummaryELMT-class}
\alias{SummaryELMT-class}
\alias{SummaryELMT}
\title{\linkS4class{SummaryELMT} class}
\description{
S4 class for a summary of \linkS4class{ELMT} objects.
}
\section{Slots}{

\describe{
\item{\code{aliased}}{A named logical vector showing if the original coefficients are
aliased.}
}}

\examples{
showClass("SummaryELMT")
}
