% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SummaryELT-class}
\alias{SummaryELT-class}
\alias{SummaryELT}
\title{\linkS4class{SummaryELT} class}
\description{
S4 class for a summary of \linkS4class{ELT} objects.
}
\section{Slots}{

\describe{
\item{\code{aliased}}{A named logical vector showing if the original coefficients are
aliased.}
}}

\examples{
showClass("SummaryELT")
}
