% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SummaryGLM-class}
\alias{SummaryGLM-class}
\alias{SummaryGLM}
\title{\linkS4class{SummaryGLM} class}
\description{
S4 class for a summary of \linkS4class{GLM} objects. It inherits from
\linkS4class{SummaryLM} class.
}
\section{Slots}{

\describe{
\item{\code{family}}{A \code{\link{family}} object used.}

\item{\code{dispersion}}{A single numeric for the estimated dispersion parameter.}
}}

\examples{
showClass("SummaryGLM")
}
