% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/print-methods.R
\docType{class}
\name{SummaryLM-class}
\alias{SummaryLM-class}
\alias{SummaryLM}
\alias{print,SummaryLM-method}
\title{\linkS4class{SummaryLM} class}
\usage{
\S4method{print}{SummaryLM}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\arguments{
\item{x}{An object of class \linkS4class{SummaryLM}.}

\item{digits}{A single integer for the number of significant digits to be
passed to \code{\link[=format]{format()}}.}

\item{signif.stars}{A single logical. If \code{TRUE}, ‘significance stars’
are printed for each coefficient.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
S4 class for a summary of \linkS4class{LM} objects.
}
\section{Slots}{

\describe{
\item{\code{statistic}}{A single numeric for the minus twice the empirical
log-likelihood ratio for the overall test of the model.}

\item{\code{df}}{A single integer for the degrees of freedom of the statistic.}

\item{\code{convergence}}{A single logical for the convergence status of the
constrained minimization.}

\item{\code{parMatrix}}{A numeric matrix of the test results of the parameters.}

\item{\code{weighted}}{A single logical for whether the given model is weighted or
not.}

\item{\code{na.action}}{Information returned by \code{\link{model.frame}} on the special
handling of \code{NA}s.}

\item{\code{call}}{Matched call.}

\item{\code{terms}}{\code{\link{terms}} object used.}

\item{\code{aliased}}{A named logical vector showing if the original coefficients are
aliased.}
}}

\examples{
showClass("SummaryLM")
}
