% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SummaryQGLM-class}
\alias{SummaryQGLM-class}
\alias{SummaryQGLM}
\title{\linkS4class{SummaryQGLM} class}
\description{
S4 class for a summary of \linkS4class{QGLM} objects. It inherits from
\linkS4class{SummaryGLM} class.
}
\examples{
showClass("SummaryQGLM")
}
