% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test2sample.R
\name{test2sample}
\alias{test2sample}
\title{Two sample test for equal mean}
\usage{
test2sample(x, y, b = 0.1, alpha = NULL, maxit = 1000, abstol = 1e-08)
}
\arguments{
\item{x}{a vector of data for one  group.}

\item{y}{a vector of data for the other  group.}

\item{b}{a momentum parameter for minimization. Defaults to .1.}

\item{alpha}{an optional step size.}

\item{maxit}{an optional value for the maximum number of iterations. Defaults to 1000.}

\item{abstol}{an optional value for the absolute convergence tolerance. Defaults to 1e-8.}
}
\description{
Two sample test for equal mean
}
\examples{
x <- rnorm(100)
y <- rnorm(100)
test2sample2(x, y)

}
