% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_ibd}
\alias{test_ibd}
\title{Empirical Likelihood Hypothesis Testing}
\usage{
test_ibd(x, c, lhs, rhs, approx = FALSE, maxit = 1000L, abstol = 1e-08)
}
\arguments{
\item{x}{a matrix of data .}

\item{c}{an incidence matrix.}

\item{lhs}{a linear hypothesis matrix.}

\item{rhs}{right-hand-side vector for hypothesis, with as many entries as rows in the hypothesis matrix.}

\item{approx}{whether to use the approximation for lambda. Defaults to FALSE.}

\item{maxit}{an optional value for the maximum number of iterations. Defaults to 1000.}

\item{abstol}{an optional value for the absolute convergence tolerance. Defaults to 1e-8.}
}
\description{
Hypothesis test for incomplete block design
}
