% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_pair}
\alias{test_pair}
\title{Two sample test for BIBD}
\usage{
test_pair(x, c, pair, maxit = 1000L, abstol = 1e-08)
}
\arguments{
\item{x}{a matrix of data .}

\item{c}{an incidence matrix.}

\item{pair}{a pair of index for parameters to be tested.}

\item{maxit}{an optional value for the maximum number of iterations. Defaults to 1000.}

\item{abstol}{an optional value for the absolute convergence tolerance. Defaults to 1e-8.}
}
\description{
Two sample test for BIBD
}
