% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/weights-methods.R
\name{weights}
\alias{weights}
\alias{weights,EL-method}
\title{Model weights}
\usage{
\S4method{weights}{EL}(object, ...)
}
\arguments{
\item{object}{An object that inherit from class \linkS4class{EL}.}

\item{...}{Not used.}
}
\description{
Extracts model weights from objects that inherit from class
\linkS4class{EL}. The weights are rescaled to up to the total number of
observations in the fitting procedure.
}
\examples{
data("airquality")
x <- airquality$Wind
w <- airquality$Day
fit <- el_mean(x, par = 10, weights = w)
weights(fit)
}
