% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clothianidin}
\alias{clothianidin}
\title{Clothianidin concentration in maize plants}
\format{
A data frame with 102 observations and 3 variables:
\itemize{
\item blk New blocks constructed from original data. The format is 'days post
planting_original block_year'.
\item trt Seed treatment.
\item clo Log transformed clothianidin concentration (µg).
}
}
\source{
Alford, Adam, and Christian H. Krupke. 2017. “Translocation of the
Neonicotinoid Seed Treatment Clothianidin in Maize.”
Edited by Michael J. Stout. PLOS ONE 12 (3): e0173836.
\doi{10.1371/journal.pone.0173836}.
}
\usage{
data("clothianidin")
}
\description{
A dataset summarizing field experiments result of seed treatments on
clothianidin concentration.
}
\details{
The original data is provided by Alford and Krupke (2017). Only some
of the shoot region observations are taken from the original data and
processed for illustration.
}
\keyword{datasets}
