% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{coef-method}
\alias{coef-method}
\title{Model coefficients}
\arguments{
\item{object}{A fitted \linkS4class{EL} object.}

\item{...}{Not used.}
}
\description{
Extracts maximum empirical likelihood estimates from a model.
}
\examples{
fit <- el_lm(formula = mpg ~ wt, data = mtcars)
coef(fit)
}
