% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/coef-methods.R
\name{coef}
\alias{coef}
\alias{coef,EL-method}
\title{Model coefficients}
\usage{
\S4method{coef}{EL}(object, ...)
}
\arguments{
\item{object}{An object that inherit from \linkS4class{EL}, including
\linkS4class{CEL}, \linkS4class{LM}, and \linkS4class{GLM}.}

\item{...}{Further arguments passed to methods.}
}
\value{
A numeric vector of the maximum empirical likelihood estimates.
}
\description{
Extracts maximum empirical likelihood estimates from a model.
}
\examples{
data("mtcars")
fit <- el_lm(mpg ~ wt, data = mtcars)
coef(fit)
}
