% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{ConfregEL-class}
\alias{ConfregEL-class}
\alias{ConfregEL}
\title{An S4 class \code{"ConfregEL"}}
\description{
An S4 class for a two-dimensional confidence region.
}
\section{Slots}{

\describe{
\item{\code{points}}{A numeric matrix with two columns for boundary points of a
confidence region.}

\item{\code{estimates}}{A numeric vector of length two for parameter estimates.}

\item{\code{level}}{A confidence level required.}

\item{\code{cv}}{A critical value for calibration of empirical likelihood ratio
statistic.}

\item{\code{pnames}}{A character vector of length two for the name of parameters.}
}}

