% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ConfregEL-class}
\alias{ConfregEL-class}
\alias{ConfregEL}
\title{\linkS4class{ConfregEL} class}
\description{
S4 class for confidence region. It inherits from \code{"matrix"}.
}
\section{Slots}{

\describe{
\item{\code{estimates}}{A numeric vector of length two for the parameter estimates.}

\item{\code{level}}{A single numeric for the confidence level required.}

\item{\code{cv}}{A single numeric for the critical value for calibration of empirical
likelihood ratio statistic.}

\item{\code{pnames}}{A character vector of length two for the name of parameters.}
}}

\examples{
showClass("ConfregEL")
}
