% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/conv-methods.R
\name{conv}
\alias{conv}
\alias{conv,EL-method}
\alias{conv,CEL-method}
\title{Convergence statistics}
\usage{
\S4method{conv}{EL}(object, ...)

\S4method{conv}{CEL}(object, ...)
}
\arguments{
\item{object}{An object that inherit from class \linkS4class{EL}, including
\linkS4class{CEL}, \linkS4class{LM}, and \linkS4class{GLM}.}

\item{...}{Further arguments passed to other methods.}
}
\description{
Extracts convergence statistics from a model.
}
\section{Methods (by class)}{
\itemize{
\item \code{EL}: Extracts the convergence status of the model with respect to
the Lagrange multiplier.

\item \code{CEL}: Extracts the convergence status of the model with respect to
the parameter.
}}

