% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{EL-class}
\alias{EL-class}
\alias{EL}
\title{\linkS4class{EL} class}
\description{
S4 class for empirical likelihood.
}
\details{
Let \eqn{X_i} be independent and identically distributed
\eqn{p}-dimensional random variable from an unknown distribution \eqn{P}
for \eqn{i = 1, \dots, n}. We assume that \eqn{P} has a positive definite
covariance matrix. For a parameter of interest
\eqn{\theta(F) \in {\rm{I\!R}}^p}, consider a \eqn{p}-dimensional smooth
estimating function \eqn{g(X_i, \theta)} with a moment condition
\deqn{\textrm{E}[g(X_i, \theta)] = 0.}
We assume that there exists an unique \eqn{\theta_0} that solves the above
equation. Given a value of \eqn{\theta}, the (profile) empirical likelihood
ratio is defined by
\deqn{R(\theta) =
  \max_{p_i}\left\{\prod_{i = 1}^n np_i :
  \sum_{i = 1}^n p_i g(X_i, \theta) = 0, p_i \geq 0, \sum_{i = 1}^n p_i = 1
  \right\}.}
The Lagrange multiplier \eqn{\lambda \equiv \lambda(\theta)} of the dual
problem leads to
\deqn{p_i = \frac{1}{n}\frac{1}{1 + \lambda^\top g(X_i, \theta)},}
where \eqn{\lambda} solves
\deqn{\frac{1}{n}\sum_{i = 1}^n \frac{g(X_i, \theta)}
  {1 + \lambda^\top g(X_i, \theta)} = 0.}
Then the empirical log-likelihood ratio is given by
\deqn{\log R(\theta) = -\sum_{i = 1}^n
  \log(1 + \lambda^\top g(X_i, \theta)).}
This problem can be efficiently solved by the Newton-Raphson method when
the zero vector is contained in the interior of the convex hull of
\eqn{\{g(X_i, \theta)\}_{i = 1}^n}.

It is known that \eqn{-2\log R(\theta_0)} converges in
distribution to \eqn{\chi^2_p}, where \eqn{\chi^2_p} has a chi-square
distribution with \eqn{p} degrees of freedom. See the references below for
more details.
}
\section{Slots}{

\describe{
\item{\code{optim}}{A list of the following optimization results:
\itemize{
\item \code{par} A numeric vector of the specified parameters.
\item \code{lambda} A numeric vector of the Lagrange multipliers of the dual
problem corresponding to \code{par}.
\item \code{iterations} A single integer for the number of iterations performed.
\item \code{convergence} A single logical for the convergence status.
\item \code{cstr} A single logical for whether constrained EL optimization is
performed or not.
}}

\item{\code{logp}}{A numeric vector of the log probabilities of the empirical
likelihood.}

\item{\code{logl}}{A single numeric of the empirical log-likelihood.}

\item{\code{loglr}}{A single numeric of the empirical log-likelihood ratio.}

\item{\code{statistic}}{A single numeric of minus twice the empirical log-likelihood
ratio with an asymptotic chi-square distribution.}

\item{\code{df}}{A single integer for the degrees of freedom of the statistic.}

\item{\code{pval}}{A single numeric for the \eqn{p}-value of the statistic.}

\item{\code{nobs}}{A single integer for the number of observations.}

\item{\code{npar}}{A single integer for the number of parameters.}

\item{\code{weights}}{A numeric vector of the re-scaled weights used for the model
fitting.}

\item{\code{coefficients}}{A numeric vector of the maximum empirical likelihood
estimates of the parameters.}

\item{\code{method}}{A single character for the method dispatch in internal
functions.}

\item{\code{data}}{A numeric matrix of the data for the model fitting.}

\item{\code{control}}{An object of class \linkS4class{ControlEL} constructed by
\code{\link[=el_control]{el_control()}}.}
}}

\examples{
showClass("EL")
}
\references{
Owen A (2001).
\emph{Empirical Likelihood}. Chapman & Hall/CRC.
\doi{10.1201/9781420036152}.

Qin J, Lawless J (1994).
``Empirical Likelihood and General Estimating Equations.''
\emph{The Annals of Statistics}, \strong{22}(1), 300--325.
\doi{10.1214/aos/1176325370}.
}
