% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/el_control.R
\name{el_control}
\alias{el_control}
\title{Control parameters for computation}
\usage{
el_control(
  maxit = 200L,
  maxit_l = 25L,
  tol = 1e-06,
  tol_l = 1e-06,
  step = NULL,
  th = NULL,
  nthreads,
  seed = sample.int(.Machine$integer.max, 1L),
  B = 10000L
)
}
\arguments{
\item{maxit}{A single integer for the maximum number of iterations for
constrained minimization of empirical likelihood. Defaults to \code{200}.}

\item{maxit_l}{A single integer for the maximum number of iterations for
evaluation of empirical likelihood. Defaults to \code{25}.}

\item{tol}{A single numeric for the convergence tolerance for the constrained
minimization. Defaults to \code{1e-06}.}

\item{tol_l}{A single numeric for the relative convergence tolerance for the
evaluation. Defaults to \code{1e-06}.}

\item{step}{A single numeric for the step size for projected gradient descent
method. Defaults to \code{NULL} and set to the reciprocal of sample size.}

\item{th}{A single numeric for the threshold for the negative empirical
log-likelihood ratio. The iteration stops if the value exceeds the
threshold. Defaults to \code{NULL} and sets the threshold to \code{200 * d}, where
\code{d} corresponds to the degrees of freedom of the limiting chi-squared
distribution of the statistic.}

\item{nthreads}{A single integer for the number of threads for parallel
computation via OpenMP (if available). Defaults to the half of the
available threads. For better performance, it is generally recommended to
limit the number of threads to the number of physical cores. Note that it
only applies to the following functions that involve multiple evaluations
or minimizations: \code{\link{confint}}, \code{\link{confreg}},
\code{\link{el_lm}}, \code{\link{el_glm}}, \code{\link{eld}}, and
\code{\link{elt}}.}

\item{seed}{A single integer for the seed for random number generation. It
only applies to \code{\link{elt}} when \code{calibrate} is set to \code{"boot"}.
Defaults to a random integer generated from 1 to the maximum integer
supported by \R on the machine, which is determined by
\code{\link[base]{set.seed}}. Only one seed is needed even when multiple
threads are used with \code{nthreads}. Each thread is given a separate seed
to produce a non-overlapping but reproducible sequence of random numbers.
The Xoshiro256+ pseudo-random number generator is used internally to work
with OpenMP.}

\item{B}{A single integer for the number of bootstrap replicates. It only
applies to \code{\link{elt}} when \code{calibrate} is set to \code{"boot"}. Defaults
to \code{10000}.}
}
\value{
An object of class of \linkS4class{ControlEL}.
}
\description{
Specifies computational details of (constrained) empirical likelihood.
}
\examples{
optcfg <- el_control(maxit = 300, th = 200, nthreads = 1)
}
\seealso{
\link{el_eval}, \link{elt}
}
