% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/el_eval.R
\name{el_eval}
\alias{el_eval}
\title{Empirical likelihood with general estimating functions}
\usage{
el_eval(g, weights = NULL, control = melt_control())
}
\arguments{
\item{g}{A numeric matrix, or an object that can be coerced to a numeric
matrix. Each row corresponds to an observation of an estimating
function.}

\item{weights}{An optional numeric vector of weights to be used in the
fitting process. Defaults to \code{NULL}, corresponding to identical
weights. If non-\code{NULL}, weighted empirical likelihood is computed.}

\item{control}{A list of control parameters set by
\code{\link{melt_control}}.}
}
\value{
A list with the following components:
\describe{
  \item{optim}{A list with the following optimization results:
    \describe{
      \item{lambda}{The Lagrange multiplier of dual problem.}
      \item{logLR}{The (weighted) empirical log-likelihood ratio value.}
      \item{iterations}{The number of iterations performed.}
      \item{convergence}{A logical vector. \code{TRUE} indicates
      convergence of the algorithm.}
    }
  }
  \item{npar}{The number of parameters.}
  \item{log.prob}{The log probabilities.}
  \item{loglik}{The log likelihood value.}
  \item{statistic}{The chi-square statistic.}
  \item{df}{The degrees of freedom of the statistic.}
  \item{p.value}{The \eqn{p}-value of the statistic.}
  \item{weights}{If non-\code{NULL} \code{weights} is supplied, the
  rescaled weights are returned.}
}
}
\description{
Computes empirical likelihood with general estimating functions.
}
\details{
\code{el_eval} evaluates empirical likelihood with a \eqn{n} by
  \eqn{p} numeric matrix argument \code{g}, whose \eqn{i}th row is
  \eqn{g(X_i, \theta)}. If \code{weights} is non \code{NULL}, the weights are
  rescaled to add up to \eqn{n}.
}
\examples{
# test for variance with known mean
x <- rnorm(100L)
sigma <- 1
g <- x^2 - sigma^2
el_eval(g)
}
\references{
Glenn, N.L., and Yichuan Zhao. 2007.
  “Weighted Empirical Likelihood Estimates and Their Robustness Properties.”
  Computational Statistics & Data Analysis 51 (10): 5130–41.
  \doi{10.1016/j.csda.2006.07.032}.

Qin, Jin, and Jerry Lawless. 1994.
  “Empirical Likelihood and General Estimating Equations.”
  The Annals of Statistics 22 (1): 300–325. \doi{10.1214/aos/1176325370}.
}
\seealso{
\link{melt_control}
}
