% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/el_glm.R
\name{el_glm}
\alias{el_glm}
\title{Fits a generalized linear model with empirical likelihood}
\usage{
el_glm(
  formula,
  family = gaussian,
  data,
  weights,
  control = list(),
  model = TRUE,
  contrasts = NULL,
  ...
)
}
\arguments{
\item{formula}{A formula object.}

\item{family}{dd}

\item{data}{A data frame containing the variables in the formula.}

\item{weights}{An optional numeric vector of weights to be used in the
fitting process. If not provided, identical weights are applied. Otherwise,
weighted empirical likelihood is computed.}

\item{control}{A list of control parameters. See ‘Details’ in
\code{\link{el_eval}}.}

\item{model}{A logical. If \code{TRUE} the model matrix used for fitting is
returned.}

\item{contrasts}{dd}

\item{...}{For glm: arguments to be used to form the default control argument
if it is not supplied directly.}
}
\value{
A list with class \code{c("el_lm", "el_test")}.
}
\description{
Fits a generalized linear model with empirical likelihood.
}
\examples{
fit <- el_lm(mpg ~ wt, mtcars)
summary(fit)
}
\references{
Owen, Art. 1991. “Empirical Likelihood for Linear Models.”
  The Annals of Statistics 19 (4).
  \doi{10.1214/aos/1176348368}.
}
\seealso{
\link{el_eval}, \link{lht}
}
