% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/el_mean.R
\name{el_mean}
\alias{el_mean}
\title{Empirical likelihood test for mean}
\usage{
el_mean(par, x, control = list())
}
\arguments{
\item{par}{Numeric vector of parameters to be tested.}

\item{x}{Numeric matrix or vector of data. For matrix \code{x}, each row corresponds to an observation.}

\item{control}{A list of control parameters. See ‘Details’.}
}
\value{
A list with class \code{"el_test"}.
}
\description{
Computes empirical likelihood for mean parameter.
}
\examples{
## scalar mean
par <- 0
x <- rnorm(100)
el_mean(par, x)

## vector mean
x <- matrix(rnorm(100), ncol = 2)
par <- c(0, 0)
el_mean(par, x)
}
\references{
Owen, Art. 1990. “Empirical Likelihood Ratio Confidence Regions.” The Annals of Statistics 18 (1). \doi{10.1214/aos/1176347494}.
}
