% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/el_mean.R
\name{el_mean}
\alias{el_mean}
\title{Empirical likelihood for the mean}
\usage{
el_mean(x, par, weights = NULL, control = el_control(), model = TRUE)
}
\arguments{
\item{x}{A numeric matrix, or an object that can be coerced to a numeric
matrix. Each row corresponds to an observation.}

\item{par}{A numeric vector of parameter values to be tested.}

\item{weights}{An optional numeric vector of weights to be used in the
fitting process. Defaults to \code{NULL}, corresponding to identical
weights. If non-\code{NULL}, weighted empirical likelihood is computed.}

\item{control}{An object of class \linkS4class{ControlEL} constructed by
\code{\link{el_control}}.}

\item{model}{A single logical. If \code{TRUE} the data matrix used for model
fitting is returned.}
}
\value{
An object of class \linkS4class{EL}.
}
\description{
Computes empirical likelihood for the mean.
}
\examples{
# scalar mean
x <- rnorm(100)
par <- 0
el_mean(x, par)

# vector mean
x <- matrix(rnorm(100), ncol = 2)
par <- c(0, 0)
el_mean(x, par)

# weighted data
x <- matrix(rnorm(100), ncol = 2)
par <- c(0, 0)
w <- rep(c(1, 2), each = 25)
el_mean(x, par, w)
}
\references{
Owen, Art. 1990. “Empirical Likelihood Ratio Confidence Regions.”
The Annals of Statistics 18 (1): 90–120. \doi{10.1214/aos/1176347494}.
}
\seealso{
\link{el_control}, \link{el_eval}, \link{elt}
}
