% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ELD-class}
\alias{ELD-class}
\alias{ELD}
\title{\linkS4class{ELD} class}
\description{
S4 class for empirical likelihood displacement. It inherits from the base
type "numeric" and has a virtual slot \code{.Data} that contains the vector of
empirical likelihood displacement values.
}
\examples{
showClass("ELD")
}
