% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ELD-class}
\alias{ELD-class}
\alias{ELD}
\title{\linkS4class{ELD} class}
\description{
S4 class for empirical likelihood displacement.
}
\section{Slots}{

\describe{
\item{\code{eld}}{A numeric vector of empirical likelihood displacement values.}
}}

\examples{
showClass("ELD")
}
