% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/print-methods.R
\docType{class}
\name{ELMT-class}
\alias{ELMT-class}
\alias{ELMT}
\alias{print,ELMT-method}
\title{S4 class \linkS4class{ELMT}}
\usage{
\S4method{print}{ELMT}(
  x,
  digits = getOption("digits"),
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\arguments{
\item{x}{An object of class \linkS4class{ELMT}.}

\item{digits}{A single integer for the number of significant digits to be
passed to \code{\link[=format]{format()}}.}

\item{signif.stars}{A single logical. If \code{TRUE}, ‘significance stars’
are printed for each coefficient.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
S4 class for empirical likelihood multiple tests.
}
\section{Slots}{

\describe{
\item{\code{alpha}}{A single numeric for the overall significance level.}

\item{\code{statistic}}{A numeric vector for the minus twice the (constrained)
empirical log-likelihood ratios.}

\item{\code{cv}}{A single numeric for the multiplicity adjusted critical value.}

\item{\code{pval}}{A numeric vector for the multiplicity adjusted \eqn{p}-values.}

\item{\code{calibrate}}{A single character for the calibration method used.}
}}

\examples{
showClass("ELMT")
}
