% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ELMT-class}
\alias{ELMT-class}
\alias{ELMT}
\title{\linkS4class{ELMT} class}
\description{
S4 class for empirical likelihood multiple tests.
}
\section{Slots}{

\describe{
\item{\code{alpha}}{A single numeric for the overall significance level.}

\item{\code{statistic}}{A numeric vector for the minus twice the (constrained)
empirical log-likelihood ratios.}

\item{\code{cv}}{A single numeric for the multiplicity adjusted critical value.}

\item{\code{pval}}{A numeric vector for the multiplicity adjusted \eqn{p}-values.}

\item{\code{calibrate}}{A single character for the calibration method used.}
}}

\examples{
showClass("ELMT")
}
