% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/elmt-methods.R
\name{elmt}
\alias{elmt}
\alias{elmt,EL-method}
\title{Empirical likelihood multiple tests}
\usage{
\S4method{elmt}{EL}(object, rhs = NULL, lhs = NULL, alpha = 0.05, control = el_control())
}
\arguments{
\item{object}{An object that inherit from class \linkS4class{EL}, including
\linkS4class{CEL}, \linkS4class{LM}, and \linkS4class{GLM}.}

\item{rhs}{A numeric vector (column matrix) or a list of numeric vectors for
the right-hand sides of hypotheses. Defaults to \code{NULL}. See ‘Details’.}

\item{lhs}{A numeric matrix or a list of numeric matrices for the left-hand
sides of hypothesis. Each row gives a
linear combination of parameters. The number of columns should be equal to
the number of parameters in \code{object}. Defaults to \code{NULL}.
See ‘Details’.}

\item{alpha}{A single numeric for the overall significance level. Defaults to
\code{0.05}.}

\item{control}{An object of class \linkS4class{ControlEL} constructed by
\code{\link[=el_control]{el_control()}}.}
}
\value{
An object of class of \linkS4class{MELT}.
}
\description{
Tests multiple linear hypotheses for objects that inherit from class
\linkS4class{EL}.
}
\examples{
n <- 100L
}
\references{
Qin, Jing, and Jerry Lawless. 1995.
“Estimating Equations, Empirical Likelihood and Constraints on Parameters.”
Canadian Journal of Statistics 23 (2): 145–59. \doi{10.2307/3315441}.
}
\seealso{
\code{\link[=el_control]{el_control()}}, \code{\link[=elt]{elt()}}
}
