% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ELT-class}
\alias{ELT-class}
\title{S4 class \linkS4class{ELT}}
\description{
S4 class for empirical likelihood test.
}
\section{Slots}{

\describe{
\item{\code{optim}}{A list with the optimization results.}

\item{\code{alpha}}{A single numeric for the minus twice the empirical
log-likelihood ratio for the overall test of the model.}

\item{\code{logl}}{A single numeric for the}

\item{\code{statistic}}{A single numeric for .}

\item{\code{cv}}{A single numeric for the critical value.}

\item{\code{pval}}{A single numeric for the \eqn{p}-value of the statistic.}

\item{\code{calibrate}}{A single character for the calibration method.}
}}

\examples{
showClass("ELT")
}
