% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getDF-methods.R
\name{getDF}
\alias{getDF}
\alias{getDF,EL-method}
\alias{getDF,logLikEL-method}
\alias{getDF,SummaryLM-method}
\title{Degrees of freedom}
\usage{
\S4method{getDF}{EL}(object)

\S4method{getDF}{logLikEL}(object)

\S4method{getDF}{SummaryLM}(object)
}
\arguments{
\item{object}{An object that inherit from \linkS4class{EL},
\linkS4class{ELT}, \linkS4class{logLikEL}, or \linkS4class{SummaryLM}.}
}
\value{
A single integer.
}
\description{
Extracts degrees of freedom from a model.
}
\examples{
data("faithful")
fit <- el_mean(faithful, par = c(3.5, 70))
getDF(fit)
}
