% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/getSigTests-methods.R
\name{getSigTests}
\alias{getSigTests}
\alias{getSigTests,LM-method}
\alias{getSigTests,SummaryLM-method}
\title{Significance tests}
\usage{
\S4method{getSigTests}{LM}(object)

\S4method{getSigTests}{SummaryLM}(object)
}
\arguments{
\item{object}{An object that inherit from \linkS4class{LM} or
\linkS4class{SummaryLM}.}
}
\value{
The form of the value returned by \code{\link[=getSigTests]{getSigTests()}} depends on the
class of its argument.
}
\description{
Extracts the results of significance tests from a model.
}
\section{Methods (by class)}{
\itemize{
\item \code{getSigTests(LM)}: Extracts a list with the optimization results of
significance tests.

\item \code{getSigTests(SummaryLM)}: Extracts a matrix with the results of significance
tests.

}}
\examples{
data("mtcars")
fit <- el_lm(mpg ~ ., data = mtcars)
getSigTests(fit)
getSigTests(summary(fit))
}
\seealso{
\code{\link[=getOptim]{getOptim()}}
}
