% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{GLM-class}
\alias{GLM-class}
\alias{GLM}
\title{\linkS4class{GLM} class}
\description{
S4 class for generalized linear models with empirical likelihood. It inherits
from \linkS4class{LM} class.
}
\examples{
showClass("GLM")
}
