% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{GLM-class}
\alias{GLM-class}
\alias{GLM}
\title{\linkS4class{GLM} class}
\description{
S4 class for generalized linear models. It inherits from \linkS4class{LM}
  class.
}
\details{
The overall test involves a constrained optimization problem. All
  the parameters except for the intercept are constrained to zero. The
  `optim` slot contains the results. When there is no intercept, all
  parameters are set to zero, and the results need to be understood in terms
  of \linkS4class{EL} class since no constrained optimization is involved.
  Once the solution is found, the log probabilities (`logp`) and the
  (constrained) empirical likelihood values (`logl`, `loglr`, `statistic`)
  readily follow, along with the degrees of freedom (`df`) and the
  \eqn{p}-value (`pval`). The significance tests for each parameter also
  involve constrained optimization problems where only one parameter is
  constrained to zero. The `sigTests` slot contains the results.
}
\section{Slots}{

\describe{
\item{\code{family}}{A [`family`] object used.}

\item{\code{dispersion}}{A single numeric for the estimated dispersion parameter.}

\item{\code{sigTests}}{A list of the following results of significance tests:
* `statistic` A numeric vector of minus twice the (constrained) empirical
log-likelihood ratios with asymptotic chi-square distributions.
* `iterations` An integer vector for the number of iterations performed for
each parameter.
* `convergence` A logical vector for the convergence status of each
parameter.}

\item{\code{call}}{A matched call.}

\item{\code{terms}}{A [`terms`] object used.}

\item{\code{misc}}{A list of various outputs obtained from the model fitting process.
They are used in other generics and methods.}

\item{\code{optim}}{A list of the following optimization results:
* `par` A numeric vector of the solution to the (constrained) optimization
problem.
* `lambda` A numeric vector of the Lagrange multipliers of the dual
problem corresponding to `par`.
* `iterations` A single integer for the number of iterations performed.
* `convergence` A single logical for the convergence status.}

\item{\code{logp}}{A numeric vector of the log probabilities of the (constrained)
empirical likelihood.}

\item{\code{logl}}{A single numeric of the (constrained) empirical log-likelihood.}

\item{\code{loglr}}{A single numeric of the (constrained) empirical log-likelihood
ratio.}

\item{\code{statistic}}{A single numeric of minus twice the (constrained) empirical
log-likelihood ratio with an asymptotic chi-square distribution.}

\item{\code{df}}{A single integer for the degrees of freedom of the statistic.}

\item{\code{pval}}{A single numeric for the \eqn{p}-value of the statistic.}

\item{\code{nobs}}{A single integer for the number of observations.}

\item{\code{npar}}{A single integer for the number of parameters.}

\item{\code{weights}}{A numeric vector of the re-scaled weights used for the model
fitting.}

\item{\code{coefficients}}{A numeric vector of the maximum empirical likelihood
estimates of the parameters.}

\item{\code{method}}{A single character for the method dispatch in internal
functions.}

\item{\code{data}}{A numeric matrix of the data for the model fitting.}
}}

\examples{
showClass("GLM")
}
