% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/el_test.R
\name{lht}
\alias{lht}
\title{Linear hypothesis test}
\usage{
lht(object, lhs, rhs, control = list())
}
\arguments{
\item{object}{A fitted "\code{el_test}" object.}

\item{lhs}{A matrix}

\item{rhs}{An optional rhs}

\item{control}{A list of control parameters. See ‘Details’.}
}
\description{
Linear hypothesis test
}
