% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lht.R
\name{lht}
\alias{lht}
\title{Linear hypothesis test}
\usage{
lht(
  object,
  rhs = NULL,
  lhs = NULL,
  alpha = 0.05,
  calibrate = c("chisq", "boot", "f"),
  control = el_control()
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{\link{lht}} was renamed to \code{\link{elt}} to avoid conflicts with
other packages.
}
\seealso{
\link{elt}
}
\keyword{internal}
