% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{LM-class}
\alias{LM-class}
\title{S4 class \linkS4class{LM}}
\description{
S4 class for linear models with empirical likelihood.
}
\section{Slots}{

\describe{
\item{\code{parTests}}{A list with the test results for each parameter:
\itemize{
\item{\code{statistic } }{Numeric vector of chi-squared statistics.}
\item{\code{convergence } }{Logical vector. \code{TRUE} indicates
convergence of the algorithm.}
}}

\item{\code{misc}}{A list with the following optimization results:
\itemize{
\item{\code{method } }{Character for method dispatch in internal
functions.}
\item{\code{par } }{Value at which empirical likelihood is minimized.}
\item{\code{lambda } }{Lagrange multiplier of the dual problem.}
\item{\code{iterations } }{Number of iterations performed.}
\item{\code{convergence } }{Convergence status.}
}}
}}

\examples{
showClass("LM")
}
