% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/logLik-methods.R
\name{logLik}
\alias{logLik}
\alias{logLik,EL-method}
\title{Empirical log-likelihood}
\usage{
\S4method{logLik}{EL}(object, ...)
}
\arguments{
\item{object}{An object that inherit from \linkS4class{EL}.}

\item{...}{Further arguments passed to methods.}
}
\value{
An object of class \linkS4class{logLikEL}.
}
\description{
Extracts empirical log-likelihood from a model evaluated at the estimated
coefficients.
}
\examples{
data("precip")
fit <- el_mean(precip, par = 40)
logLik(fit)
}
