% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/logLik-methods.R
\name{logLik}
\alias{logLik}
\alias{logLik,EL-method}
\title{Empirical log-likelihood}
\usage{
\S4method{logLik}{EL}(object, ...)
}
\arguments{
\item{object}{A fitted \linkS4class{EL} object.}

\item{...}{Not used.}
}
\value{
An object of class \linkS4class{logLikEL}.
}
\description{
Extracts empirical log-likelihood from a model evaluated at the estimated
  coefficients.
}
\examples{
fit <- el_lm(formula = mpg ~ wt, data = mtcars)
logLik(fit)
}
