% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{logLikEL-class}
\alias{logLikEL-class}
\title{S4 class \linkS4class{logLikEL}}
\description{
S4 class for empirical log-likelihood.
}
\section{Slots}{

\describe{
\item{\code{logLik}}{Empirical log-likelihood.}

\item{\code{df}}{Degrees of freedom or the number of (estimated) parameters in the
model.}
}}

\examples{
showClass("logLikEL")
}
