% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{logLikEL-class}
\alias{logLikEL-class}
\alias{logLikEL}
\title{\linkS4class{logLikEL} class}
\description{
S4 class for empirical log-likelihood.
}
\section{Slots}{

\describe{
\item{\code{logLik}}{A single numeric for the empirical log-likelihood.}

\item{\code{df}}{A single integer for the degrees of freedom or the number of
(estimated) parameters in the model.}
}}

\examples{
showClass("logLikEL")
}
