% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/logLR-methods.R
\name{logLR}
\alias{logLR}
\alias{logLR,EL-method}
\alias{logLR,ELT-method}
\title{Empirical log-likelihood ratio}
\usage{
\S4method{logLR}{EL}(object, ...)

\S4method{logLR}{ELT}(object, ...)
}
\arguments{
\item{object}{An object that inherit from \linkS4class{EL} or
\linkS4class{ELT}.}

\item{...}{Further arguments passed to methods.}
}
\value{
A single numeric.
}
\description{
Extracts empirical log-likelihood ratio from a model.
}
\examples{
data("precip")
fit <- el_mean(precip, par = 40)
logLR(fit)
}
