% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt-package.R
\docType{package}
\name{melt-package}
\alias{melt}
\alias{melt-package}
\title{melt: Multiple Empirical Likelihood Tests}
\description{
Performs multiple empirical likelihood tests for linear and generalized linear models. The core computational routines are implemented using the 'Eigen' C++ library and 'RcppEigen' interface, with OpenMP for parallel computation. Details of multiple testing procedures are given in Kim, MacEachern, and Peruggia (2021) \href{https://arxiv.org/abs/2112.09206}{arXiv:2112.09206}. This work was supported by the U.S. National Science Foundation under Grants No. SES-1921523 and DMS-2015552.
}
\references{
Kim E, MacEachern S, Peruggia M (2021).
“Empirical Likelihood for the Analysis of Experimental Designs.”
arxiv:2112.09206. URL \url{https://arxiv.org/abs/2112.09206}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/markean/melt}
  \item \url{https://markean.github.io/melt/}
  \item Report bugs at \url{https://github.com/markean/melt/issues}
}

}
\author{
\strong{Maintainer}: Eunseop Kim \email{kim.7302@osu.edu}

Other contributors:
\itemize{
  \item Steven MacEachern [contributor, thesis advisor]
  \item Mario Peruggia [contributor, thesis advisor]
}

}
\keyword{internal}
