% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/nobs-methods.R
\name{nobs}
\alias{nobs}
\alias{nobs,EL-method}
\alias{nobs,SummaryEL-method}
\alias{nobs,SummaryLM-method}
\title{Number of observations in a model}
\usage{
\S4method{nobs}{EL}(object, ...)

\S4method{nobs}{SummaryEL}(object, ...)

\S4method{nobs}{SummaryLM}(object, ...)
}
\arguments{
\item{object}{An object that contains the number of observations.}

\item{...}{Further arguments passed to methods.}
}
\value{
A single integer.
}
\description{
Extracts the number of observations from a model.
}
\examples{
data("precip")
fit <- el_mean(precip, par = 40)
nobs(fit)
}
\seealso{
\linkS4class{EL}
}
