% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SD-class}
\alias{SD-class}
\alias{SD}
\title{\linkS4class{SD} class}
\description{
S4 class for standard deviation. It inherits from \linkS4class{EL} class.
}
\examples{
showClass("SD")
}
