% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SummaryLM-class}
\alias{SummaryLM-class}
\title{S4 class \linkS4class{SummaryLM}}
\description{
S4 class for a summary of \linkS4class{LM} objects.
}
\section{Slots}{

\describe{
\item{\code{statistic}}{Minus twice the constrained empirical log-likelihood ratio
for the overall test of the model.}

\item{\code{df}}{Degrees of freedom of the statistic.}

\item{\code{convergence}}{Convergence status of the minimization.}

\item{\code{parMatrix}}{A numeric matrix of the test results of the parameters.}

\item{\code{weighted}}{A logical for whether the given model is weighted or not.}

\item{\code{na.action}}{Information returned by \code{\link[stats]{model.frame}} on
the special handling of NAs.}

\item{\code{call}}{Matched call.}

\item{\code{terms}}{\code{\link[stats]{terms}} object used.}

\item{\code{aliased}}{Named logical vector showing if the original coefficients are
aliased.}
}}

\examples{
showClass("SummaryLM")
}
